pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==2.3.3",
    "flask-cors==4.0.0",
    "google-generativeai==0.3.2",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
]
